🔨 Remove old pinned version of package
A long time ago we pinned a spesific version of the `sharp` library
because of bugs on arm-processors, since then over a year have lapsed
and that problem should no longer be present in new versinos of the
library.

TODO: in the future we should also go over and update gatsby and plugins
to the newest version. I don't think that its related to current build
problems but it will come and bite us in the ass at some point.